Fix bad hash functor in test driver
As noticed by Bill, one of the hash functors used by the test
drivers returns 'bool' rather than 'size_t', and in fact always
returns 'true', so that many of our test cases are storing all
elements in a single bucket.

It is disturbing that there is no QoI test of the distribution
of our elements across buckets in the common cases, a missing
concern that should be addressed in a future update to the
test driver - once we have an appropriate metric to test
effective distribution of buckets at varying load factors.