Extra check to catch hash table corruption
The added check is for the kind of issue that should be reported by
test drivers rather than redundant checking in class methods - it is
specifically looking for bugs in an implementation, rather than misuse.
This should speed issue discovery during the test process though, and
will be removed for the final release.