Start testing HashTable with function pointers
Initial test covers only the first bootstrap test case for
the basic constructor and primary manipulators.  The is some
concern at the size of the test driver generating the product
of some test scenerios, rather than the sum, so we may want
to optimize the test matrix a little before generalizing
this test configuration for all test cases.