Doc updates describing when HashTable may throw
In order to address some awkward scenarios with large containers,
or extreme values for 'maxLoadFactor', we are allowing HashTable
to throw 'std::length_error' in some circumstances.  This update
provides documentation, but not yet a full implementation, of
the new contracts.

It also cleans up and documents the functor-wrapper classes used
to handle user-supplied functors with function call operators
lacking a 'const' qualification.