Restore the string-hashing behavior of hash<const char *>
While we updated the default hash functor used by the STLport hash
containers to use a specific functor to hash strings, we forgot to
account for users making explicit use of 'bsl::hash', and live code
breaks if 'bsl::hash<const char *>' starts hashing the pointer
value, rather than the assumed string value.  This patch restores
the hashing-as-string behavior, with exactly the same poor hash
function that was traditionally used, and writes a message to the
console the first time that it is called.  The hash-as-string
behavior is guarded by 'BDE_OMIT_INTERNAL_DEPRECATED' macros,
so should not escape into the open source distribution.