bsl::hash should use the correct deprecation macros
Correct the use of the OMIT_TRANSITIONAL macro, replaced by
OMIT_INTERNAL_DEPRECATED.

Also, add a minimal test case to verify the expected behavior of the
hash function on 'const char *' pointers, in either build mode.  This
quickly flagged the error that the function in the .cpp file was
defined as 'inline', so was not being written to the .o file - also
now fixed.