Fix bsls_stopwatch.t.cpp to compile with C++11
The basic problem with this test driver is that it is using 'std'
headers that it should not depend on.  Those headers, on gcc in
C++11, are transitively including further standard headers, and
consequently introduced an ambiguity on the name 'array'.

The fix is to entirely remove the 'std' dependency and replace
its use with the 'bsls_bsltestutil' component, designed for
this specific purpose.