Framework for testing non-BDE allocators with HashTable
This change-set formalises the idea of testing with standard
conforming non-BDE allocators.  It initially supports the
stateless 'bsltf::StdTestAllocator', and will shortly be
extended to support a stateful version, with a variety of
propagation traits.

The key to making this work within a single test case is to
arrange for the existing 'config' test allocators to be used
in some way by the non-BDE allocators as an implementation
detail, and proivide a way to expose that details so that the
test driver can obtain the address of the relevant test
allocator.  This allows the regular BDE memory-testing
methodology to be employed.