Initial support for testing stateful allocators
Here we add a new allocator component to 'bsltf' that can be used
to verify types that claim to support C++11 standard allocators.
This component, 'bsltf_stdstatefulallocator', is based on the
existing stateless allocator component, 'bsltf_stdtestallocator'.
The key difference is that the stateful allocator wraps a
'bslma::TestAllocator' pointer, which serves as its state, and
that this component targets the minimal set of C++11 allocator
requirements, rather than the C++03 requirements.  This means
adding configurable support for the range of allocator
propagation traits, although they are not used by this component
itself.

As the test driver and usage example of this component are still
pillaged from the stateless allocator test component, a little
work remains to make this truly robust, and there are a couple
of methods that need better documentation.  However, the
basic component itself should be fundamentally sound, and can
now be used experimentally to verify the final set of test
concerns bootstrapping the HashTable component.

Once the configuration with the new test allocator component
is confirmed, we will take the final step back to make this
component rigorous.