HashTable bootstrap tests stateful STL allocators
This patch takes advantage of the new bsltf component to support
testing of containers with C++11 allocators having varied
propagation properties.  At this point, test case 2 is almost
complete, barring some general clean-up and better doc of the
new type-bases concerns that have been tested, and the primite
testing of 'maxLoadFactor'.

One important validation that remains to be added is the test for
properly passing on the object allocator to the container
elements, when the allocator is 'bsl::allocator' and the
elements have the 'bslma::UsesBslmaAllocator' trait.  Once that
final test is checked in, we can finally proceed to test case 3.