unbreak nightly tests for incomplete test driver
The test driver that was copied as the basis for testing
the new test-allocator for stateful allocators was written
for stateless allocators, and so tested some wrong
assumptions for the new component.  This hasty patch
corrects the assumptions and provides a little more
testing in this area, but this is far from a rigorous
implementation of the test driver itself.