Resolve static analysis warning on Microsoft
The Microsoft standard library does not annotate 'abort'
as a non-returning function, and this raises warning in
static analysis that a couple of functions in the test
driver have code paths that can return without supplying
a return value.  These warnings are treated as errors in
the BDE nightly build scripts, so we add an additional
call to 'return', only for Microsoft compiles, to satisfy
the static analyser.