Attempt to work around IBM-specific bug
The IBM AIX platform (only) is complaining that when we
attempt to create a stateless std-test allocator, the
address of the allocator we are requesting is not that
of the currently installed delegate allocator, nor is
it the new-delete allocator singleton.  We suspect that
this platform is somehow producing multiple copies of a
static variable (a known problem in the past) so attempt
to work around by installing a known global test allocator
in 'main', and ensure that the static global std-test
allocator is declared and defined only in that single
translation unit of the test library, and not in a header
where the compiler may see it and think to create multiple
copies.