Resolve HashTable testing bug on AIX
The bug related to how we set and reset the global 'state'
allocator used for testing stateless standard allocators.
For some reason, this genuine bug in the test driver was
reported only by AIX test runners.  I suspect that the
other compilers must be re-using stack space claimed by
loops for variables declared subsequent to the loop, and
we got (un)lucky in picking a compatible address for the
new test allocator installed at the end of the test case,
when poking into the far corner cases.

This code could probably be simplified further, now that
the real cause has been identified, knowing that it is
not some phantom bug where IBM have, in the past, been
known to have multiple definitions for static objects.

The final remaining aspect of test case 2 is to confirm
the scoped behavior of allocators, where the HashTable
elements should use the same allocator as the HashTable
itself, if the allocator is 'bsl::allocator' and the
element type has the 'bslma::UsesBslmaAllocator' trait.