More on HashTable test case 3
This simple patch broadens the range of configurations tested
in test case 3, which currently tests only the generator
functions 'gg' and 'ggg'.  Ulitmately, this function should
be testing the whole range of test support machinery, but
there will likely be massive redundancy if every configuration
tests every facet of test machinery - much of that machinery
is shared with identical template instantiations between the
different configurations.  Some thought should be applied to
rolling out broader support for testing the testing infra-
structure of this test driver.