Expand type-coverage of HashTable test 4
This patch updates test case 4 using the new allocator and functors
framework for testing in this test driver, but loses the ability to
set the state of the stateful functors until the next update.
Given we have extensive test history that the HashTable template
works correctly with basic stateful functors, we are happy to get
a day or so running with the broader type-based coverage while the
framework is extended to transparently support attempting to set
and read the state of generalized functor - not exclusively the
small set of known stateful ones.