Extend test-type coverage to copy constructor
Extended the basic copy constructor test to support the
range of types tested in test cases 2,3 and 4.  We still
lack proper confirmation that the state of stateful
functors is copied (although I am confident this is the
case).  The case of stateful allocators *is* handled,
although oddly, the case of *stateless* allocators is
not tested, as our framework around that case raises
false errors - due to some contracts written for test
components that assumed we would not try to create two
stateless allocators from objects with different state.

As per test case 2, we are still not verifying that
the elements allocated for a container copy are
allocated from the same allocator as the container
itself, when the elements have the 'UsesBslmaAllocator'
trait, and the container allocator type is bsl::allocator.