Fix HashTable::setMaxLoadFactor for empty tables
The 'setMaxLoadFactor' function of 'bslstl::HashTable' was calling
the function to compute the suggested number of buckets to support
the growth strategy with the current container size, which might
potentially be 0, which would be out of contract.  This was showing
up as an assertion in safe builds.

Correct the bug, to always pass a mimimum of at least one element,
and document the reason for this narrow contract in the .cpp file.