Reduce cost of testing undefined behavior in HashTable swap
While our definition of 'swap' for any container is that it
has undefined behavior unless the allocators are the same,
we also provide an (undocumented) expensive implementation
that makes two copies with the correct allocators, and swaps
each container with its respective copy.

The test that this undefined behavior worked correctly, and
was strongly exception safe, was the dominant cost in running
test case 8, which already has two pairs of nested loop-pairs
iterating over load factors, and numbers of elements.  As we
do not guarnatee the undefined behavior, we settle for merely
testing one quite expensive case for each configuration of
the test case, rather than providing the full incremental
test of exception safety on each innermost loop iteration.

This should now allow us to test with a much wider variety of
test case configurations, for a similar runtime cost.  It
should be noted that the compile-time cost of this test is
already unreasonable, and we must filter the tested range of
configurations sooner rather than later.