Revised test plan for HashTable
Revise the plan for testing HashTable after the initial
ten value-semantic test cases.  The key issue is that in
order to properly test the default constructor, we will
want to call some mutating operations, such as insert,
to verify that the container constructed as empty in this
manner has exactly the same behavior as the empty table
constructed back in test case 2.  As an empty HashTable
initially has no buckets, this means we must call the
real 'insert' function rather than the test driver shim
used to validate the first ten test cases.  This, in
turn, is going to require that we validate the rehash
functions, that become the new test case 11.  We will
validate 'insert' and the default constructor together
in test case 12, and then proceed to validate the rest
of the public interface.

In addition to reorganizing the test plan, this patch
rolls out additional test-type coverage for some of
the existing proto-test-cases, including a basic test
for 'rehashForNumBuckets'.  This turned up a minor
concern in how we deduce the number of buckets to
grow to in a rehash, preferring to mandate that if we
go to the effort of a rehash, there will always be
capacity to insert on more element before another
rehash is need - although an awkward user may still
fiddle with the 'maxLoadFactor' to force an earlier
rehash.