Reduce runtime cost of HashTable tests
Drop the number of load factors tested from five
to three.  This array is looped over in a number
of tests, and nested in at least on test bringing
the variations down from 25 to 9.

We may want to restore the breadth of testing, or
pick better edge values, when the driver is
complete, but this should reduce the nightly
burden a little over the next few days.

Unfortunately, it does little for the compile
time cost, which is still excessive, particularly
on AIX.