Catch upexpected 'bad_alloc' exceptions in test driver
The HashTable test driver has, in its breathing test, a simple
request to set a 'maxLoadFactor' of '9e-9f' for a container of
a single element.  This expression was not guarded by a try/
catch block, and it appears that it will throw a 'bad_alloc'
exception on at least two of our test platforms, AIX and
Windows.

This patch puts the appropriate try/catch in place, in the
appropriate build modes, and clarifies the flow of the test
of 'rehashForNumBuckets' a little.  Still need to test the
second 'rehashFor...' function, and to test exception safety,
and to verify basic exception safety when the user-supplied
functors throw in the middle of a rehash.