Reorganize HashTable test driver to simpliy compilation
In order to reduce some complexity for the compiler, especially
in optimiziating builds, separate out each test case into its
own test function.  Noramlly, simply using the template test
facility is sufficient for this, as there are only a couple
of invokations per test case.  As the primary template under
test in this driver has so many variations to test though,
even these cases are becoming large and unwieldy simply
dispatching to the actual testing code.  Therefore, we
separate out the dispatch functions, which often expand to
200 lines or more after preprocessing, to their own
distinct functions.