Simplify HashTable breathing test
There is no need to maintain the breathing test as a template
test case, as we are happy to test a single, specific instantiation
of 'HashTable' and call each of its members.  In fact, specifically
testing a single instance, rather than writing and testing more
generic code, lends an extra test that we are not simply fooling
ourselves thinking that all the template code is magically
expanding in some way to just give us the answers we want, without
actually executing any code.