Reduce HashTable test driver complexity for IBM
This patch simply reduces test coverage on IBM in order to
reduce the complexity of the test driver so that it can
compile.  We retain full coverage on other platforms.  Once
the test driver is complete, we are likely to revisit the
question of how many instantiations of each test template
are necessary for full coverage, reducing the complexity
on all platforms, and we shall revisit the question of IBM
coverage again at that point.