Broaden type coverage for HashTable operator= tests
This patch broadens the range of types tested by test
case 9, which is already the most expensive test case.
Informal testing on Linux suggests that the test can
still complete within the allotted time, although we
know from the nightly regression test runners that it
frequently is responsible for timeouts on a number of
platforms.