Reorganize HashTable test driver for easier documentation
Filter all the declarations to the top of the test driver,
just under the usage example, where it become much easier
to see the range of test support infrastructure available
in this test driver, and to document it correctly.  This
also helps pass the various code insspection tools deployed
by the BDE team.

Note that this patch is fundamentally about the reorganization
of the file, and the actual documentation will be filled in
progressively over the next sequence of patches, as further
testing concerns are addressed.