Make OSX builds default to Clang for compilation.
The GCC included in current versions of Xcode is quite old and has a
number of bugs that make it difficult to use for building BSL. Since
Xcode includes Clang and that is the supported compiler for the OSX
environment (and is used when the build is performed by xcodebuild),
this patch changes the build system to use Clang on OSX instead of GCC
for Makefile-driven builds. Users can still override this behavior if
they wish by specifying "CC=gcc CXX=g++" on the 'make' command line.