Finish reorganization of HashTable test driver
This patch identifies the last of the classes and class
templates that need to be properly ordered and documented,
and provides some initial documentation in most cases.

The contract for the 'evil' boolean type now includes
(in)equality comparison operators that, in turn,
return evil boolean values.  There is otherwise
minimal functional change, merely moving member function
definitions out of line from the class definitions
themselves, which provides for easier documentation and
checklisting of tests for the test machinery classes.