Separate new test components from HashTable
The hash table test driver has a lot of generic test machinery.
This patch extracts some of the more useful elements as 'bsltf'
components in their own right.

There is also a first pass at the first 'insert' function
test case.  This mostly copies the bootstrap test, but
creates HashTables with a minimal number of buckets, that
will then grow according to the container's growth
strategy.  This relies on the 'rehash' functions being
adequately tested in case 12.