Add a reserve feature to HashTable
The four unordered containers each have a 'reserve' method
that performs a rehash to guarnatee there are enough buckets
to grow the container without further rehashes.  The BDE
implementation of these containers is also employing a node
pool, and it would make sense to reserve that number of
nodes in the pool, as well as growing the bucket array.

This patch renames the 'rehashForNumElements' method of
HashTable to 'reserveForNumElements', and pre-emptively
calls a 'reserve' on its node pool before attempting
to rehash.  The hash table documentation is updated,
and a missing pre-condition in the node pool contract
was highlighted, and also documented.