Further cleanup of HashTable test driver
This patch eliminates one of the redundant hash functors used
to test HashTable, as the StatefulHash testing concerns are
adequately handled by one of the other functors.

Correct the two functors providing 'operator==' as a member
function rather than as a free function - it did not need
private access to the class.

Plus some minor formatting issues.