HashTable support for non-copyable functors
The C++11 standard merely requires that the comparator and hasher functors
are merely default constructible, for the default constructor.  Using a
functor that is *only* default constructible will produce a container that
cannot be copied, but should function perfectly well within the scope of
a single function.  This patch addresses several issues when testing with
such types.