Memory test patch for unordered_set.t.cpp
The problem with the test driver is that it correctly checked
whether the default allocator was being used unexpectedly, but
forgot to allow for the bsltf::TestValuesArray component using
the default allocator to construct its elements.  This is an
ongoing issue with that component that should be addressed,
but in the meantime we simply move the construction of the
test value array above the installation of the default allocator
for the test.