Add a 'Profile' configuration and code coverage tools.
This patch adds a 'Profile' configuration to common.gypi, and when the build
is done using GCC, adds specific compiler and linker flags to cause the library
and test drivers to be instrumented for profiling (and code coverage). In addition
it adds the 'tools/bsl_coverage' script which constructs a Makefile that uses 'lcov'
to gather and aggregate code coverage information, and it adds the 'lcov_cobertura'
tool which converts the lcov-style output into Cobertura-style output.