Prepare better contracts for bsltf::TestValuesArray
This test-facility component does not clearly document its full
purpose as a recyclable single-pass test value sequence.  This
patch clarfifies some of the contracts, notably for iterator
invalidation, and brings some parts of the component closer
to standard conformance, such as the iterator that claims to
be an input iterator.

The next patch should properly eliminate accidental use of
the default allocator, which will require a new contract
of the 'CONVERTER' policy, and an update to the test
facility 'create' function, to support a 'createInplace'
contract too.