Ongoing TestValueArray work
This patch should eliminate all use of the default allocator by
the bsltf::TestValuesArray facility.  However, if we stop
installing default allocator guards around code that was previously
using this as a workaround, we expose additional uses of the default
allocator that may need to be accounted for separately.  This is
particularly noticeable in the HashTable test driver that tests
support for non-BDE allocators, which do not respect the trait
bslma::UsesBslmaAllocator - it is meaningless to a non-BDE allocator
and so the default allocator will be used when inserting types
that use BDE allocators into such containers.

In order to avoid breaking the overnight tests, while getting
test coverage of the updated facility, this patch retains all
the existing default-allocator guards.  This detail will be
addressed tomorrow.