Improve some HashTable.t.cpp memory usage estimates
Now that 'bsltf::TestValuesArray' no longer uses the default allocator
to create a temporary populating each element of the array, we can
remove an un-necessary default allocator guard from the 'ggg' function
and get a better picture of unexpected usage.

THe main issue that needs addressing is that the non-BDE allocators
that we use for testing do not support the scoped allocator model
where the container passes its allocator to each element in their
constructor.  Therefore, when testing the non-BDE allocators with
an object allocator other than the default allocator, the memory
usage arithmetic should allow for the default allocator being
used to construct each element in the container, if those elements
in turn need an allocator (and not otherwise).