Cleanup some excess DefaultAlloctorGuards
Now that the TestValueArray component does not use the default
allocator, we can remove the workaround from test drivers that
install a default allocator guard in the 'ggg' function while
they create their TestValueArray.  This patch safely removes
the guard from associative and unordered 'set'-like components,
and notes that there appear to be some outstanding issues to
resolve for the 'map'-like components.  The guards remain on
the 'map'-like test drivers while the issues are investigated
further.