bsltf workaround for MSVC
It looks like Microsoft cannot handle pointer-to
-member types in bslalg_scalarprimitives.  Will
need to investigate this further in 'bslalg' test
drivers later, but for now this workaround appears
to solve the problem for 'bsltf'.  Currently the
workaround is enabled unless we define a macro with
a crazy but descriptive name.  I will go back and
remove this original version of the code once I
verify that the Microsoft workaround, using simple
assignment to initialize the result buffer, does
not break on other platforms.  If that turns out
to be an issue, we will simply rename the crazy
macro as a platform-detect macro for the MSVC
compiler.