Resolve segfault with map TestValueArrays
Switching to the new model of populating bsltf::TestValueArray
objects so that the default allocator is not used was causing
some odd problems in the bslstl_map test driver, and similar
test drivers for map-like, but not set-like, containers.

After some investigation it turns out that the problem is
passing a 'bsls::ObjectBuffer.buffer()' call straight to
the 'emplace' function.  The return type of 'buffer' is
a 'char *', rather than a 'TYPE *', and so emplace will
create a 'char' object rather than the expected value within
the buffer.  This bug lies in the test driver, and has been
replicated in all the map-like test drivers.

Note that this patch has a lot of diagnostic code lurkin
behind pre-processor macros, and will be cleaned up in the
next pass - but I wanted to be sure a copy with all the
diagnostic machinery available exists in the git tree.