Simplify HashTable test driver for slow platforms
This patch simply expands on the idea that we cannot broadly test
all combinations of all type concerns on all test drivers on all
platforms - something has to give.  Given that IBM and Sun are
proving the most problematic, this patch simply disabled most
of the template instantations by guarding them with #ifdef blocks,
which will be revisitted once the test driver is complete, and we
can pay attention to simplifying the test coverage to a more
minimlistic set of tests to cover the type-based concerns.