Catch a couple of operator& issues
A couple of tests were relying on a publicly accessible
'operator&' for arbitrary type parameters, which fail
when instantiated with our most awkward type on Windows.

For the latest set of changes, we have declared the
operator as private in our test type, which makes it
much simpler to find and resolve errors.  Unfortunately,
we currently support types abusing 'operator&' only on
Windows, so testing for this case is useful only on that
platform too.