Work around Sun optimized exception-free build issue
The Sun compiler is crashing trying to compile the unordered_map
test driver for optimized builds, but only when exceptions are
disabled.

This patch resolves a number of issues with exception free builds,
before simply adding a computed configuration macro to that one
test driver to detect the Sun/no-exceptions/optimized combination,
and elide the usage example in that case.

We suspect the cause may be the optimizer trying to do something
clever with a large ammount of test in this case, as that seems
to be the distinguishing feature of this one troublesome test
driver.  However, we have not reduced an example to file file
a ticket with Sun.