Resolve MSVC warnings in header
MSVC does not like intializing data members that are pointers
with the value of 'this' in a constructor.  However, there
does not appear to be a safe coding idiom to switch to that
Microsoft are happy with to disable the warning, so we must
specifically detect the platform and deploy pragmas to
disable the warning locally.  We do not want to disable the
warning globally, in case we miss a genuinely valuable case
of the warning catching unexpected behavior.  However, in
this case we are simply storing the address of this object
to validate as an invariant in the destructor, a specific
feature of this class, so this is indeed our correct and
expected behavior.