Disable unordered_map usage example on IBM/SAFE
There is some odd interaction of IBM SAFE builds with the inliner
that cause the usage example to segfault inside the first call to
'partial_sort'.  The segfault appears to come from an inlined
char_traits function trying to assign the trailing '\0' to a
string representation, possible for an empty string, in a vector
of 'pair<string, int>' objects (being sorted).  Given the error
is very specific to just this function, in just this test driver,
and only in SAFE builds, we are presming (without further proof)
that this is an IBM optimizer bug that has not yet been narrowed.
Therefore, the immediate response is to disable this test case,
much as we have to do for Sun optimized builds.  It is disturbing
that this one usage example is turning up (apparent) compiler bugs
in two different platforms, for seemingly unrelated causes, but
that is life.