Ongoing HashTable work
This patch addresses a number of issues.

Most importantly, it fixes some calls to undefined behavior in
'bsl::allocator', where a large number of buckets might be
requested, but not so large that the low-load factor tests
flag an invalid (not computable) length.

We also nail down the 'swap' contract, so that it is always
well defined, although potentially expensive, to call the
free function swap, and the member function 'swap' has a
narrow contract that requires either the same allocator for
both hash tables, or that the allocators propagate on swap.

Finally there is some ongoing work polishing the formal
testing concerns and plans, although this is still far
from complete.  It is known that the formatting of this
part of the test driver is currently non-standard, and
shall be rectified shortly.  However, we need to commit
now in order to get into the overnight regression tests.