Allocator related patches
This commit checks that attempts to construct a HashTable in such
a way that the amount of memory required cannot be computed throws
the exoected 'length_error' exception, while simply requesting
more memory than may realistically be available throws a 'bad_alloc'
exception.

In working through the details of this patch it was noted by code
inspection that there was a subtle bug in the HashTable
implementation that had not been revealed simply because all the
allocators we have used in test drivers happen to have a
nested 'size_type' typedef that is an alias for 'std::size_t'.
If it should happen to alias a different type, as is permitted
for arbitrary standard-conforming allocators, the class would
fail to compile.

In order to help catch such issues in the future, the simplest
of the 'bsltf' std-test allocators explicitly sets these typedefs
to 'unsigned int' and 'int'.  We know that these are sufficient
to provide a memory for any test driver allocator, and likewise
may well differ from the exact native size types for 32 bit
platforms, even if they have the same representation, and will
almost certainly be both a different type and representation
on 64-bit platforms.  This is sufficient to exercise the code
paths that we are concerned about.

After updating this test allocator, a number of test drivers
promptly failed to compile, notably the associative an
unordered containers test drivers, that made use of the
updated allocator in a specific test case for STL allocators.
This was not an issue for the HashTable test driver, as it
integrates STL and bsl allocators into its test pattern for
each test case.  Rather than revise all these test drivers
in a manner similar to the HashTable test driver, the minimal
fix was deployed, of using a slightly different instantiation
of the test harness class for just that one test case.