Resolve numerous allocator-related warnings
The addition yesterday of a testing allocator whose size_type
is not 'std::size_t' has triggered an avalanche of warnings in
the nightly test builds.  For good measure, most of these
warnings appear only when compiling with gcc 4.7.2 on the Sun
platform, they are not present for Linux builds with the same
compiler.

This is an interim patch - further warning cleanup is still
required, as is a review of the most appropriate way to
resolve many of these warnings.  It is suspected that in
more that a few cases, the preferred fix may be to change
a variable or member's type, rather than resort to casts.
Likewise, where we are casting, it might be prudent to have
some bounds-tests to validate that we are not losing out
to numeric overflows.