HashTable support for functors taking non-const refs
The definition of a predicate in C++ allows for functions and functors
that take their arguments by non-const reference, as long as calling
that functionality does not actually change the state of a passed
argument or call a non-const qualified function not under a similar
guarnatee (e.g., 'begin').  This patch is quite messy, as it remains
full of #if guarded code in case we back out this work that is
required for conformace, but is a real edge case to support.

at this point, the bootstrap, primary manipulators and accessors
pass with the new configuration.  There remains a const-qualification
detail in the implementation of 'operator==' that is preventing the
higher test cases from compiling with such functors for now.  This
seems a reasonable point to check into integration and see the
current behavior compiles and runs correctly across our whole range
of compilers, before spending further effort in this direction.

For the sake of sanity, this and further work is occuring on a
new branch, 'unordered_testing_revised', branched off from the
previous 'unordered_doc_and_testing' branch at the point this
work on the funny functors started.