Finish integrating HashTable testing for awkward functors
This completes the initial support for HashTables instantiated with
functors taking arguments by non-const reference, as required for
full conformance with the C++ standard.  This is tested only as
well as the current state of the test driver, but is supported for
all current test cases.  It is not expected that we can support
the entire public API for these functors, just as we do not support
all calls for all value types, e.g., we do not support 'operator=='
if the ValueType does not support 'operator=='.  In particular, the
public 'find' function will not be supported for functors that
accept their arguments by non-const reference, which does somewhat
restrict the usefulness of such types.