HashTable support for references and function-types
This patch provides basic support for the callable template parameter
types to be references, including function references, const-qualified
functors, or even plain function types.  Test coverage runs as far as
test case 4, but it is known that these policies will not support all
of the higher level test cases as the containers will be neither copyable
nor swappable.