'Fix' bsltf test alloctor test driver
This patch persuades the bsltf::StdTestAllocator test driver to
pass in the nightly builds.  For now the patch retains the old
code block for comparision while we confirm that we have the
intended new semantic.  At some point the testing concerns should
be restated, the test updated, and the old code blocks removed.