HashTable test driver for platforms without ref-collapsing
The C++ support for 'reference collapsing' just missed the 2003 standard,
it was among the first defect reports applied after that standard was
published.  While many of our compilers support this feature out of the
box, some of our older of pricklier compilers still insist on the letter
of the 2003 standard.

Some parts of the HashTable test driver rely on the reference collapsing
rules when trying to test support for reference-types as either of the
hasher or key-comparator template parameters.  While it is thought that
the HashTable itself has been corrected for this issue, the test driver
makes more pervasive use and will be difficult to correct cleanly given
all the other constraints on the various testing functions.

As a workaround, we introduce a platform-check macro, local to just the
.t.cpp file, that diagnoses platforms that to not support reference
collapsing.  With this feature-detect macro, we disable that tests that
would otherwise depend upon this feature.