Robustify HashTable generator tests
It turns out the sample data thrown at the 'gg' and 'ggg'
functions in test case 3 were nowhere near as rigorous as
I had thought, still being the initial data copied across
from the regular 'set' container test.  Adding rigor to
the data set turned up some issues in the test machinery
that are in the process of being addressed.

The positive is that so far, all the issues highlighted
turned out to be issues in the test driver and not the
component under test.  It is a little worrying that the
test driver has been running without reporting any of
these issues for so long.