Patch for IBM testing stateful hasher/comparators
The IBM compiler seems to have some partial ordering issues resolvimg
ambiguity of the function tempalte taking generic 'const &' arguments
and the more specialized template taking pointer-to-function-of-two
arguments.  Passing the latter by 'const &' shakes out any remaining
ambiguity.

There is a further issue that the 'comparator' and 'hasher' member
functions do not resolve (again on IBM) when the template parameter
is a reference type.  This is worked around simply by disabling that
one configuration on IBM for test cases that require this function.