Clearer organization of HashTable test driver
Reordered parts of this test driver, so that the template-based
concerns are closer to the configurations designed to address
those concerns, and machinery to build the template testing
framework is pulled up into a more easily separable part of the
file, so that it might be extracted as a distinct component
in the future.

In order to ease following the testing, the test case functions
are now listed numerically, so that the sense of testing the
component reads easily, in the linear direction of simply reading
the file.  This is possible only because the 'swith' statement in
main immediately dispatches to subsidiary test functions, as this
'switch' must maintain the reverse ordering if it is to support the
notion of test case zero.

As part of the reorganization, it became obvious where we had
duplicate test coverage, missing/inconsistent test coverage,
or testing of unsupported features such as propagating allocators
that are not detected by 'bsl::allocator_traits'.  This has now
been rationalized, with much of the excess removed, and needed
tests restored.